backend_vbb/requirements.txt
fastapi
uvicorn[standard]
requests
pymongo[srv]
dnspython
python-dotenv
redis

